+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Sprint 3 (November) [20 Nov, 2023 - 30 Nov, 2023]
 ### [released]
 - Implement Password difficulty bar in while type password in sign up
+- Change brand name 'bachelor' to 'House Rent'
 
 ## Sprint 2 (November) [10 Nov, 2023 - 20 Nov, 2023]
 ### [released]